Moved SES aut-configuration and properties to io.awspring.cloud.v3.autoconfigure.ses package for readability.
