+++ v3/spring-cloud-aws-v3-autoconfigure/src/main/java/io/awspring/cloud/v3/autoconfigure/ses/properties/AwsSesProperties.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.awspring.cloud.v3.autoconfigure.properties;
+package io.awspring.cloud.v3.autoconfigure.ses.properties;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
