+++ v3/spring-cloud-aws-v3-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,5 +1,5 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 io.awspring.cloud.v3.autoconfigure.CredentialsProviderAutoConfiguration,\
 io.awspring.cloud.v3.autoconfigure.RegionProviderAutoConfiguration,\
-io.awspring.cloud.v3.autoconfigure.SesAutoConfiguration
+io.awspring.cloud.v3.autoconfigure.ses.SesAutoConfiguration
