+++ v3/spring-cloud-aws-v3-autoconfigure/src/test/java/io/awspring/cloud/v3/autoconfigure/SesAutoConfigurationTest.java
@@ -16,6 +16,7 @@
 
 package io.awspring.cloud.v3.autoconfigure;
 
+import io.awspring.cloud.v3.autoconfigure.ses.SesAutoConfiguration;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.boot.autoconfigure.AutoConfigurations;
